Code/requirements.txt
pyFluidSynth
basic-pitch
tensorflow
pretty_midi
soundfile